Refactor getting mdcontent from Module.md
